change common reference Signed-off-by: GavinXu520 <meiyan532484710@163.com>
